voice-system/backend/requirements.txt
Flask>=2.3.0
google-generativeai==0.5.4
google-cloud-texttospeech==2.16.2
flask-cors==4.0.0
